move to a util file so that I can import it from Jupyter
